requirements.txt
absl-py==2.4.0
accelerate==0.29.0
aiohappyeyeballs==2.6.1
aiohttp==3.13.5
aiosignal==1.4.0
annotated-doc==0.0.4
annotated-types==0.7.0
anyascii==0.3.3
anyio==4.13.0
attrs==26.1.0
audioread==3.1.0
certifi==2026.2.25
cffi==2.0.0
charset-normalizer==3.4.7
click==8.3.2
contourpy==1.3.3
coqpit-config==0.2.5
coqui-tts==0.27.5
coqui-tts-trainer==0.3.3
cycler==0.12.1
decorator==5.2.1
dnspython==2.8.0
docopt==0.6.2
einops==0.8.2
email-validator==2.3.0
fastapi==0.111.0
fastapi-cli==0.0.24
filelock==3.25.2
fonttools==4.62.1
frozenlist==1.8.0
fsspec==2026.2.0
grpcio==1.80.0
h11==0.16.0
hf-xet==1.4.3
httpcore==1.0.9
httptools==0.7.1
httpx==0.28.1
huggingface_hub==0.36.2
idna==3.11
inflect==7.5.0
Jinja2==3.1.6
joblib==1.5.3
kiwisolver==1.5.0
ko-speech-tools==0.1.0
lazy-loader==0.5
librosa==0.11.0
llvmlite==0.47.0
Markdown==3.10.2
markdown-it-py==4.0.0
MarkupSafe==3.0.3
matplotlib==3.10.8
mdurl==0.1.2
monotonic-alignment-search==0.2.1
more-itertools==11.0.2
mpmath==1.3.0
msgpack==1.1.2
multidict==6.7.1
networkx==3.6.1
num2words==0.5.14
numba==0.65.0
numpy==1.26.4
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.20.5
nvidia-nvjitlink-cu12==12.9.86
nvidia-nvtx-cu12==12.1.105
orjson==3.11.8
packaging==26.1
pillow==12.2.0
platformdirs==4.9.6
pooch==1.9.0
propcache==0.4.1
protobuf==7.34.1
psutil==7.2.2
pycparser==3.0
pydantic==2.13.2
pydantic_core==2.46.2
pydub==0.25.1
Pygments==2.20.0
pyparsing==3.3.2
pysbd==0.3.4
python-dateutil==2.9.0.post0
python-dotenv==1.2.2
python-multipart==0.0.9
PyYAML==6.0.3
regex==2026.4.4
requests==2.33.1
rich==15.0.0
rich-toolkit==0.19.7
safetensors==0.7.0
scikit-learn==1.8.0
scipy==1.17.1
setuptools==82.0.1
shellingham==1.5.4
six==1.17.0
soundfile==0.12.1
soxr==1.0.0
starlette==0.37.2
sympy==1.14.0
tensorboard==2.20.0
tensorboard-data-server==0.7.2
threadpoolctl==3.6.0
tokenizers==0.22.2
torch==2.4.0+cu121
torchaudio==2.4.0+cu121
tqdm==4.67.3
transformers==4.57.0
triton==3.0.0
typeguard==4.5.1
typer==0.24.1
typing-inspection==0.4.2
typing_extensions==4.15.0
ujson==5.12.0
urllib3==2.6.3
uvicorn==0.29.0
uvloop==0.22.1
watchfiles==1.1.1
websockets==16.0
Werkzeug==3.1.8
yarl==1.23.0
